#278: Move SMTP Server Rule Class into EmailSenderIT
